+++ .vscode/settings.json
@@ -31,5 +31,6 @@
   ],
   "[yaml]": {
     "editor.formatOnSave": false
-  }
+  },
+  "prettier.enable": false
 }
